virtio-pci: add SR-IOV capability
This enables SR-IOV emulation on virtio-pci devices by adding SR-IOV capability
It also introduces a newly added property 'sriov_max_vfs'
to enable or disable the SR-IOV feature on the virtio-pci device in guest,
as well as to specify the maximum number of VFs that can be created in the guest.
Currently only virtio-net is supported.
Also, the vendor ID and device ID remain the same for both the PF and VF,
enabling existing guest PF drivers to be used in the VF without any modifications.

Signed-off-by: Yui Washizu <yui.washidu@gmail.com>
Message-Id: <1689731808-3009-2-git-send-email-yui.washidu@gmail.com>